third_party/blink/web_tests/http/tests/devtools/console-show-all-messages-expected.txt

Tests that console shows messages only from specific context when show target checkbox is checked.


Running: testInitialState
message from page!
message from iframe!

Running: testPageOnlyMessages
message from page!

Running: testFrameOnlyMessages
message from iframe!

Running: testAllMessagesWithFrameContext
message from page!
message from iframe!
